Py3 (#7)
* Gradually converting the Python programs to Python 3.

* ignore the backup files from 2to3 conversion
these are now python 3 programs

* Besides the usual 2to3 conversions also convert the screen position
computations to use integer (//) division.

* Mostly routine 2to3 changes. Items to note:
The conversion from bytes to unicode needs to be managed explicitly now.
Ascii poems did not require any thought as to when to convert to unicode.
I preserved the global except statements, but I would urge you to get rid
of them. They should never be used casually because of the risk of "handling"
an exception that is totally unexpected and not really getting handled.

* Removed the bytes.encode('utf8'). This proved to be an error when I added an Italian poem.
The file was read as a stream of bytes, so there was nothing to be done.
The file is UTF-8, so the decoding by the client application works correctly.

* Added an Italian poem to better test the unicode handling.
Most changes relate to sending bytes over the network connections
and converting the byte streams to unicode strings.
Made changes to error handling to see errors that were getting suppressed.
Imported the implementer decorator to replace implements from zope.interfaces

* Update README

* Ususal changes from 2to3-3.5 updating print, etc.
Usual changes to use bytes for network data.

* Update README

Co-authored-by: Dave Peticolas <dave@krondo.com>

* removed the comments in fastpoetry_plugin that referenced the obsolete implements function.

changed the error handling in the transformedpoetry server. I had changed the server to
send error messages in place of poems when errors occurred. That violates the protocol.
Now the server logs all of its errors. That should be sufficient to debug client and server
errors in the future.

* Should test before push. I introduced a syntax error in my except statement in my previous
commit. That is now fixed and I log the error that triggered the exception.

* I had never run this failure program. I may have skipped it as unnecessary for
my learning.
The only changes are the automatic ones from 2to3.

* I think this is the last file I failed to convert.
The changes are the usual 2to3 conversions plus bytes for network streams
and conversion to unicode strings for finished poems.

* Removed unnecessary comments about bytes.
Changed variable name bytes to chunks. This avoids masking the bytes function/type.

Co-authored-by: Dave Peticolas <dave@krondo.com>